server/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
Pillow==9.1.0
psycopg2==2.9.3
python-decouple==3.6